+++ src/love/states/menu/titleMenu.lua
@@ -28,16 +28,24 @@ local canMove = true
 
 return {
 	enter = function(self, previous)
+		local awaiting = save.save.awaitingExploitation
+		if awaiting then
+			awaitingExploitation = true
+			save.save.awaitingExploitation = false
+			save.writeSave()
+		end
 		canMove = true
         selectSound = paths.sound('menu/select')
         confirmSound = paths.sound('menu/confirm')
-		logo = graphics:newAnimatedSprite('dave/title/logoBumpin', {{name = 'logo bumpin', anim = 'idle', loops = true}}, 'idle', true)
+		logo = graphics:newAnimatedSprite('dave/title/logoBumpin'..(awaitingExploitation and 'expunged' or ''), {{name = 'logo bumpin', anim = 'idle', loops = true}}, 'idle', true)
 		logo.x, logo.y = -350, -125
 		logo.sizeX, logo.sizeY = 1.25, 1.25
 
-		girlfriendTitle = graphics:newAnimatedSprite('dave/title/gfDanceTitle', {{name = 'gfDance', anim = 'idle', loops = true}}, 'idle', false, nil, {smartOffsets = true})
-		girlfriendTitle.x, girlfriendTitle.y = 325, 65
-		girlfriendTitle.sizeX, girlfriendTitle.sizeY = 1.25, 1.25
+		if not awaitingExploitation then
+			girlfriendTitle = graphics:newAnimatedSprite('dave/title/gfDanceTitle', {{name = 'gfDance', anim = 'idle', loops = true}}, 'idle', false, nil, {smartOffsets = true})
+			girlfriendTitle.x, girlfriendTitle.y = 325, 65
+			girlfriendTitle.sizeX, girlfriendTitle.sizeY = 1.25, 1.25
+		end
 
 		titleEnter = graphics:newAnimatedSprite('dave/titleEnter', {{name = 'Press Enter to Begin', anim = "anim"}, {name = 'ENTER PRESSED', anim = 'pressed'}}, 'anim')
 		titleEnter.x, titleEnter.y = 225, 350
@@ -51,9 +59,9 @@ return {
 
 		--switchMenu(1)
 
-		music = paths.music('menu/menu')
+		music = paths.music(awaitingExploitation and 'menu/evilmenu' or 'menu/menu')
 		music:setLooping(true)
-		--music:play()
+		music:play()
 	end,
 
 	musicStop = function(self)
@@ -62,8 +70,11 @@ return {
 
 	update = function(self, dt)
 		--print(love.timer.getTime())
-		girlfriendTitle:update(dt)
+		if not awaitingExploitation then
+			girlfriendTitle:update(dt)
+		end
 		titleEnter:update(dt)
+		logo:update(dt)
 		--titleEnter:animate("anim", true)
 
 		if not drawTransition and canMove then
@@ -77,24 +88,20 @@ return {
 					switchState(menuSelect)
 				end)
 			end
-
-			logo:update(dt)
 		end
 	end,
 
 	draw = function(self)
 		love.graphics.push()
-			love.graphics.translate(graphics.getWidth() / 2, graphics.getHeight() / 2)
+		love.graphics.translate(graphics.getWidth() / 2, graphics.getHeight() / 2)
 
-			love.graphics.push()
-				love.graphics.scale(cam.sizeX, cam.sizeY)
+		logo:draw()
 
-				logo:draw()
+		if not awaitingExploitation then
+			girlfriendTitle:draw()
+		end
+		titleEnter:draw()
 
-				girlfriendTitle:draw()
-				titleEnter:draw()
-
-				love.graphics.pop()
 		love.graphics.pop()
 	end,