+++ CHANGELOG.md
@@ -19,10 +19,9 @@ note: this version uses a new version of love.elf, so donwload it from the new s
 this fixed one crash oops
 
 # v0.2.0 (wip)
-
 ### Gameplay
 - Character dances are 100% on beat now
-
+- You can now exit the game at any time by holding start and select at the same time
 ### Porting
 - The console (pressing 7 is now replaced by holding all the shoulders and triggers and pressing select)
 - Five Night's at Muko's (the whole game, inside this port, do the 7 thing on kabunga im not lying)